+++ pyborehole/borehole.py
@@ -1,15 +1,14 @@
 import pandas as pd
-import numpy as np
-import shapely
-from shapely.geometry import Point, LineString
+from shapely.geometry import Point
 from pyproj import CRS
 import pyproj
-import matplotlib.pyplot as plt
-from typing import Union, List, Tuple
+from typing import Union, Tuple
 import geopandas as gpd
 
 from pyborehole.deviation import Deviation
-from pyborehole.well_design import WellDesign
+from pyborehole.design import WellDesign
+from pyborehole.logs import LASLogs, DLISLogs
+from pyborehole.litholog import LithoLog
 
 
 class Borehole:
@@ -955,7 +954,7 @@ class Borehole:
                 raise ValueError('Provided columns are not within the DataFrame')
 
         # Create deviation
-        self.deviation = Deviation(self,
+        self.deviation = Deviation(borehole=self,
                                    path=path,
                                    delimiter=delimiter,
                                    step=step,
@@ -1132,612 +1131,22 @@ class Borehole:
         self.df.loc['Litholog', 'Value'] = self.has_litholog
 
     def add_well_design(self):
+        """Add well design object to borehole
 
-        self.well_design = WellDesign()
+        Examples
+        ________
+            >>> borehole.add_well_design()
+            >>> borehole.well_design
+            Pipes: {}
+            Cements: {}
 
-
-class WellTops(Borehole):
-    """Class to initiate Well Tops.
-
-    Parameters
-    __________
-        path : str
-            Path to the well tops, e.g. ``path='Well_Tops.csv'``.
-        delimiter : str
-            Delimiter to read the well tops file correctly, e.g. ``delimiter=','``.
-        unit : str
-            Unit of the depth measurements, e.g. ``unit='m'``.
-
-    """
-
-    def __init__(self,
-                 path: str,
-                 delimiter: str = ',',
-                 unit: str = 'm'):
-
-        # Checking that the path is of type str
-        if not isinstance(path, str):
-            raise TypeError('The path must be provided as string')
-
-        # Checking that the delimiter is of type str
-        if not isinstance(delimiter, str):
-            raise TypeError('The delimiter must be provided as string')
-
-        # Checking that the unit is provided as string
-        if not isinstance(unit, str):
-            raise TypeError('The unit must be provided as string')
-
-        self.df = pd.read_csv(path, delimiter=delimiter)
-
-        self.df['Unit'] = unit
-
-
-class LithoLog(Borehole):
-    """Class to initiate the LithoLog.
-
-    Parameters
-    __________
-        path : str
-            Path to the litholog, e.g. ``path='LithoLog.csv'``.
-        delimiter : str
-            Delimiter to read the litholog file correctly, e.g. ``delimiter=','``.
-
-    """
-
-    def __init__(self,
-                 path: str,
-                 delimiter: str = ',', ):
-
-        # Checking that the path is of type str
-        if not isinstance(path, str):
-            raise TypeError('The path must be provided as string')
-
-        # Checking that the delimiter is of type str
-        if not isinstance(delimiter, str):
-            raise TypeError('The delimiter must be provided as string')
-
-        self.df = pd.read_csv(path, delimiter=delimiter)
-
-
-class LASLogs(Borehole):
-    """Class to initiate a Well Log Object.
-
-    Parameters
-    __________
-        path : str
-            Path to the well logs, e.g. ``path='logs.las'``.
-
-    """
-
-    def __init__(self,
-                 borehole,
-                 path: str):
-
-        # Importing lasio
-        try:
-            import lasio
-        except ModuleNotFoundError:
-            ModuleNotFoundError('lasio package not installed')
-
-        # Opening LAS file
-        las = lasio.read(path)
-
-        # Extracting DataFrame from LAS file
-        self.df = las.df()
-
-        # Creating DataFrame from curve data
-        self.curves = pd.DataFrame(list(zip([las.curves[i]['original_mnemonic'] for i in range(len(las.curves))],
-                                            [las.curves[i]['mnemonic'] for i in range(len(las.curves))],
-                                            [las.curves[i]['descr'] for i in range(len(las.curves))],
-                                            [las.curves[i]['unit'] for i in range(len(las.curves))])),
-                                   columns=['original_mnemonic',
-                                            'mnemonic',
-                                            'descr',
-                                            'unit'])
-
-        # Creating DataFrame from well header
-        self.well_header = pd.DataFrame(list(zip([las.well[i]['mnemonic'] for i in range(len(las.well))],
-                                                 [las.well[i]['unit'] for i in range(len(las.well))],
-                                                 [las.well[i]['value'] for i in range(len(las.well))],
-                                                 [las.well[i]['descr'] for i in range(len(las.well))])),
-                                        columns=['mnemonic',
-                                                 'unit',
-                                                 'value',
-                                                 'descr'])
-
-        # Creating DataFrame from parameters
-        self.params = pd.DataFrame(list(zip([las.params[i]['mnemonic'] for i in range(len(las.params))],
-                                            [las.params[i]['unit'] for i in range(len(las.params))],
-                                            [las.params[i]['value'] for i in range(len(las.params))],
-                                            [las.params[i]['descr'] for i in range(len(las.params))])),
-                                   columns=['mnemonic',
-                                            'unit',
-                                            'value',
-                                            'descr'])
-
-        self.well_tops = borehole.well_tops
-
-    def plot_well_logs(self,
-                       tracks: Union[str, list],
-                       depth_column: str = 'MD',
-                       colors: Union[str, list] = None,
-                       add_well_tops: bool = False,
-                       fill_between: int = None):
-        """Plot well logs.
-
-        Parameters
-        __________
-
-        tracks : Union[str, list]
-            Name/s of the logs to be plotted, e.g. ``tracks='SGR'`` or ``tracks=['SGR', 'K'].
-        depth_column : str
-            Name of the column holding the depths, e.g. ``depth_column='MD'``.
-        colors : Union[str, list]
-            Colors of the logs, e.g. ``colors='black'`` or ``colors=['black', 'blue'].
-        add_well_tops : bool, default = False
-            Boolean to add well tops to the plot.
-        """
-        # Selecting tracks
-        df = self.df[tracks].reset_index()
-
-        if isinstance(tracks, str):
-            # Creating plot
-            fig, ax = plt.subplots(1, 1, figsize=(1 * 2, 8))
-
-            ax.plot(df[tracks], df[depth_column], color=colors)
-            ax.grid()
-            ax.invert_yaxis()
-            buffer = (max(df[depth_column]) - min(df[depth_column])) / 20
-            ax.set_ylim(max(df[depth_column]) + buffer, min(df[depth_column]) - buffer)
-            ax.tick_params(top=True, labeltop=True, bottom=False, labelbottom=False)
-            ax.xaxis.set_label_position('top')
-            ax.set_xlabel(tracks + ' [%s]' %
-                          self.curves[self.curves['original_mnemonic'] == tracks].reset_index(drop=True)['unit'].iloc[
-                              0], color='black')
-            ax.set_ylabel(depth_column + ' [m]')
-
-            if fill_between:
-                left_col_value = np.min(df[tracks].dropna().values)
-                right_col_value = np.max(df[tracks].dropna().values)
-                span = abs(left_col_value - right_col_value)
-                cmap = plt.get_cmap('hot_r')
-                color_index = np.arange(left_col_value, right_col_value, span / 100)
-                # loop through each value in the color_index
-                for index in sorted(color_index):
-                    index_value = (index - left_col_value) / span
-                    color = cmap(index_value)  # obtain color for color index value
-                    ax.fill_betweenx(df[depth_column], df[tracks], left_col_value, where=df[tracks] >= index,
-                                     color=color)
-
-            return fig, ax
-
-        elif isinstance(tracks, list):
-
-            if add_well_tops:
-                j = 1
-            else:
-                j = 0
-
-            if not colors:
-                colors = [None] * len(tracks)
-
-            # Creating plot
-            fig, ax = plt.subplots(1,
-                                   len(tracks) + j,
-                                   figsize=(len(tracks) * 1.8, 8),
-                                   sharey=True)
-
-            # Helping variable for adding well tops
-            if add_well_tops:
-                for index, row in self.well_tops.df.iterrows():
-                    ax[0].axhline(row[self.well_tops.df.columns[1]], 0, 1, color='black')
-                    ax[0].text(0.05, row[self.well_tops.df.columns[1]] - 1, s=row[self.well_tops.df.columns[0]],
-                               fontsize=6)
-                    ax[0].grid()
-                    ax[0].axes.get_xaxis().set_ticks([])
-
-            # Plotting tracks
-            for i in range(len(tracks)):
-                ax[i + j].plot(df[tracks[i]], df[depth_column], color=colors[i])
-                ax[i + j].grid()
-                ax[i + j].invert_yaxis()
-                buffer = (max(df[depth_column]) - min(df[depth_column])) / 20
-                ax[i + j].set_ylim(max(df[depth_column]) + buffer, min(df[depth_column]) - buffer)
-                ax[i + j].tick_params(top=True, labeltop=True, bottom=False, labelbottom=False)
-                ax[i + j].xaxis.set_label_position('top')
-                ax[i + j].set_xlabel(tracks[i] + ' [%s]' %
-                                     self.curves[self.curves['original_mnemonic'] == tracks[i]].reset_index(drop=True)[
-                                         'unit'].iloc[0],
-                                     color='black' if isinstance(colors[i], type(None)) else colors[i])
-                ax[0].set_ylabel(depth_column + ' [m]')
-
-            if fill_between is not None:
-                left_col_value = np.min(df[tracks[fill_between]].dropna().values)
-                right_col_value = np.max(df[tracks[fill_between]].dropna().values)
-                span = abs(left_col_value - right_col_value)
-                cmap = plt.get_cmap('hot_r')
-                color_index = np.arange(left_col_value, right_col_value, span / 100)
-                # loop through each value in the color_index
-                for index in sorted(color_index):
-                    index_value = (index - left_col_value) / span
-                    color = cmap(index_value)  # obtain color for color index value
-                    ax[fill_between + j].fill_betweenx(df[depth_column], df[tracks[fill_between]], left_col_value,
-                                                       where=df[tracks[fill_between]] >= index,
-                                                       color=color)
-
-            plt.tight_layout()
-
-            return fig, ax
-
-    def plot_well_log_along_path(self,
-                                 log: str,
-                                 coordinates: pd.DataFrame,
-                                 spacing: float = 0.5,
-                                 radius_factor: float = 75):
+        .. versionadded:: 0.0.1
         """
 
-        Parameters
-        __________
+        # Setting has well design variable
+        self.has_well_design
 
+        # Adding well design
+        self.well_design = WellDesign(borehole=self)
 
-        Raises
-        ______
-            ModuleNotFoundError
-                Raises error if PyVista is not installed.
-            TypeError
-                If the wrong input data types are provided.
-            ValueError
-                Raises error if the wrong column names are provided.
 
-        """
-        # Importing pyvista
-        try:
-            import pyvista as pv
-        except ModuleNotFoundError:
-            ModuleNotFoundError('PyVista package not installed')
-
-        if not {'Northing', 'Easting', 'True Vertical Depth Below Sea Level'}.issubset(coordinates.columns):
-            raise ValueError('The coordinates DataFrame must contain a northing, easting and true vertical depth '
-                             'below sea level column')
-
-        coordinates = coordinates[['Easting', 'Northing', 'True Vertical Depth Below Sea Level']].to_numpy()
-
-        logs = self.df.reset_index()[['MD', log]]
-
-        points = resample_between_well_deviation_points(coordinates=coordinates,
-                                                        spacing=spacing)
-
-        # polyline_well_path = polyline_from_points(points=coordinates)
-
-        polyline_well_path_resampled = pv.Spline(points)
-
-        points_along_spline = get_points_along_spline(spline=polyline_well_path_resampled,
-                                                      dist=logs['MD'].values)
-
-        polyline_along_spline = polyline_from_points(points=points_along_spline)
-
-        polyline_along_spline['values'] = logs[log].values
-
-        tube_along_spline = polyline_along_spline.tube(scalars='values',
-                                                       radius_factor=radius_factor)
-
-        return tube_along_spline
-
-
-class DLISLogs(Borehole):
-    """Class to initiate a Well Log Object.
-
-        Parameters
-        __________
-            path : str
-                Path to the well logs, e.g. ``path='logs.dlis'``.
-
-        """
-
-    def __init__(self,
-                 borehole,
-                 path: str,
-                 nodata: Union[int, float] = -9999):
-
-        # Importing dlisio
-        try:
-            from dlisio import dlis
-        except ModuleNotFoundError:
-            ModuleNotFoundError('dlisio package not installed')
-
-        # Opening DLIS file
-        dlis, *tail = dlis.load(path)
-
-        # Getting column names
-        columns = [channel.name for channel in dlis.channels]
-
-        # Getting Curves
-        curves = [channel.curves() for channel in dlis.channels]
-
-        # Creating DataFrame from curves
-        df = pd.DataFrame(curves).T
-
-        # Assigning column names
-        df.columns = columns
-
-        # Replace NaN Values
-        df = df.replace(nodata, np.NaN)
-
-        # Extracting DataFrame from LAS file
-        self.df = df
-
-
-def resample_between_well_deviation_points(coordinates: np.ndarray,
-                                           spacing: Union[float, int]) -> np.ndarray:
-    """Resample between points that define the path of a well.
-
-    Parameters
-    __________
-        coordinates: np.ndarray
-            Nx3 Numpy array containing the X, Y, and Z coordinates that define the path of a well.
-
-    Returns
-    _______
-         points_resampled: np.ndarray
-            Resampled points along a well.
-
-    .. versionadded:: 1.0.x
-
-    """
-
-    # Checking that the coordinates are provided as np.ndarray
-    if not isinstance(coordinates, np.ndarray):
-        raise TypeError('Coordinates must be provided as NumPy Array')
-
-    # Checking that three coordinates are provided for each point
-    if coordinates.shape[1] != 3:
-        raise ValueError('Three coordinates X, Y, and Z must be provided for each point')
-
-        # Creating list for storing points
-    list_points = []
-
-    # Iterating over points and creating additional points between all other points
-    for i in range(len(coordinates) - 1):
-        dist = np.linalg.norm(coordinates[i] - coordinates[i + 1])
-        num_points = int(dist // spacing)
-        points = np.linspace(coordinates[i], coordinates[i + 1], num_points + 1)
-        list_points.append(points)
-
-    # Converting lists of points into np.ndarray
-    points_resampled = np.array([item for sublist in list_points for item in sublist])
-
-    return points_resampled
-
-
-def polyline_from_points(points: np.ndarray):
-    """Create PyVista PolyLine from points
-
-    Parameters
-    __________
-
-        points: np.ndarray
-            Points defining the PolyLine.
-
-    Return
-    ______
-        poly: pv.core.pointset.PolyData
-
-    .. versionadded:: 1.0.x
-
-    """
-
-    # Importing pyvista
-    try:
-        import pyvista as pv
-    except ModuleNotFoundError:
-        ModuleNotFoundError('PyVista package not installed')
-
-    # Checking that the points are of type PolyData Pointset
-    if not isinstance(points, np.ndarray):
-        raise TypeError('The points must be provided as NumPy Array')
-
-    # Creating PolyData Object
-    poly = pv.PolyData()
-
-    # Assigning points
-    poly.points = points
-
-    # Creating line values
-    the_cell = np.arange(0, len(points), dtype=np.int_)
-    the_cell = np.insert(the_cell, 0, len(points))
-
-    # Assigning values to PolyData
-    poly.lines = the_cell
-
-    return poly
-
-
-def get_points_along_spline(spline,
-                            dist: np.ndarray):
-    """Return the closest point on the spline a given a length along a spline.
-
-    Parameters
-    __________
-
-        spline: pv.core.pointset.PolyData
-            Spline with the resampled vertices
-
-        dist: np.ndarray
-            np.ndarray containing the measured depths (MD) of values along the well path
-
-    Returns
-    _______
-
-        spline.points[idx_list]: pv.core.pyvista_ndarray.pyvista_ndarray
-            PyVista Array containing the selected points
-
-    .. versionadded:: 1.0.x
-
-    """
-
-    # Importing pyvista
-    try:
-        import pyvista as pv
-    except ModuleNotFoundError:
-        ModuleNotFoundError('PyVista package not installed')
-
-    # Checking that the spline is a PyVista PolyData Pointset
-    if not isinstance(spline, pv.core.pointset.PolyData):
-        raise TypeError('The well path/the spline must be provided as PyVista PolyData Pointset')
-
-    # Checking that the distances are provided as np.ndarray
-    if not isinstance(dist, np.ndarray):
-        raise TypeError('The distances must be provided as np.ndarray')
-
-    # Creating list for storing indices
-    idx_list = []
-
-    # Getting index of spline that match with a measured value and append index to list of indices
-    for distance in dist:
-        idx = np.argmin(np.abs(spline.point_data['arc_length'] - distance))
-        idx_list.append(idx)
-
-    points = spline.points[idx_list]
-
-    return points
-
-
-def resample_log(log: Union[gpd.GeoDataFrame, LineString, pd.DataFrame],
-                 resampling: int,
-                 column_name: str = None,
-                 resampling_start=0,
-                 resampling_end=None,
-                 rounding_precision=5,
-                 drop_first: bool = False,
-                 drop_last: bool = False,
-                 ) -> gpd.GeoDataFrame:
-    """Resample one log. Data must be provided as GeoDataFrame with the log Data as Shapely Points.
-
-
-
-    """
-    # Creating gdf from values
-    if isinstance(log, pd.DataFrame):
-        log = gpd.GeoDataFrame(geometry=[LineString(log[[column_name, 'DEPT']].values)])
-    # Extracting LineString from Value and assigning column name
-    if isinstance(log, gpd.GeoDataFrame):
-        if not column_name:
-            column_name = 'X'
-        log = log.geometry.iloc[0]
-    # Assigning column name
-    elif isinstance(log, shapely.geometry.LineString):
-        if not column_name:
-            column_name = 'X'
-
-    # Extracting coordinates
-    x = [coords[0] for coords in list(log.coords)]
-    y = [coords[1] for coords in list(log.coords)]
-
-    # Getting last position for resampling
-    if not resampling_end:
-        resampling_end = np.negative(np.floor(y[-1]), where=False)
-
-    # Create LineStrings along Log
-    linestrings = [LineString([(min(x),
-                                y_sample),
-                               (max(x),
-                                y_sample)]) for y_sample in np.arange(resampling_start,
-                                                                      resampling_end,
-                                                                      -resampling)]
-
-    # Create intersection points between LineStrings and log
-    points = [
-        shapely.wkt.loads(shapely.wkt.dumps(log.intersection(line_intersect), rounding_precision=rounding_precision))
-        for line_intersect in linestrings if not log.intersection(line_intersect).is_empty]
-
-    # Creating GeoDataFrame from log
-    gdf_resampled = gpd.GeoDataFrame(geometry=points)
-
-    # Setting columns
-    gdf_resampled[column_name] = gdf_resampled.geometry.centroid.x
-    gdf_resampled['Y'] = np.round(gdf_resampled.geometry.centroid.y, 1)
-
-    # Concatenating GeoDataFrames
-    gdf_resampled = pd.concat([pd.DataFrame.from_dict(
-        {'geometry': Point(x[0], y[0]), column_name: [x[0]], 'Y': [y[0]]}, orient='columns'),
-        gdf_resampled,
-        pd.DataFrame.from_dict(
-            {'geometry': Point(x[-1], y[-1]), column_name: [x[-1]], 'Y': [y[-1]]},
-            orient='columns')
-    ], ignore_index=True)
-
-    # Dropping duplicates and resetting index
-    gdf_resampled = gdf_resampled.drop_duplicates().reset_index(drop=True)
-
-    # Dropping first depth
-    if drop_first:
-        gdf_resampled = gdf_resampled[1:].reset_index(drop=True)
-
-    # Dropping last depth
-    if drop_last:
-        gdf_resampled = gdf_resampled[:-1]
-
-    return gdf_resampled
-
-
-def resample_logs(logs: pd.DataFrame,
-                  resampling: int,
-                  resampling_start=0,
-                  resampling_end=None,
-                  rounding_precision=5,
-                  drop_first: bool = True,
-                  drop_last: bool = True, ):
-    # Resampling DataFrames
-    dfs = [resample_log(log=logs[['DEPT', column]],
-                        resampling=resampling,
-                        column_name=column,
-                        resampling_start=resampling_start,
-                        resampling_end=resampling_end,
-                        rounding_precision=rounding_precision,
-                        drop_first=drop_first,
-                        drop_last=drop_last) for column in logs.columns.drop('DEPT')]
-
-    # Concatenating DataFrames
-    df = pd.concat(dfs, axis=1).drop('geometry', axis=1)
-
-    # Dropping duplicate columns
-    df = df.loc[:, ~df.columns.duplicated()]
-
-    return df
-
-
-def merge_logs(paths: List[str],
-               resampling: Union[float, int]) -> pd.DataFrame:
-    try:
-        import lasio
-    except ModuleNotFoundError:
-        ModuleNotFoundError('lasio package not installed')
-
-    # Opening LAS Files as DataFrames
-    dfs = [lasio.read(path).df().reset_index() for path in paths]
-
-    # Resampling logs
-    dfs_resampled = [resample_logs(logs=df,
-                                   resampling=-resampling,
-                                   drop_first=True,
-                                   drop_last=True) for df in dfs]
-
-    # Getting minimum und maximum depth
-    miny = [df.describe().loc['min']['Y'] for df in dfs_resampled]
-    maxy = [df.describe().loc['max']['Y'] for df in dfs_resampled]
-
-    # Creating DataFrame ranging across the entire depth range
-    df_depth = pd.DataFrame(np.arange(min(miny), max(maxy) + 1, 1), columns=['Y'])
-
-    # Copying depth DataFrame
-    merged_df = df_depth.copy(deep=True)
-
-    # Merging DataFrames
-    for i in range(len(dfs_resampled)):
-        dfs_resampled[i] = dfs_resampled[i].rename(columns={'X': 'X_%s' % i})
-        merged_df = pd.merge(merged_df, dfs_resampled[i], on='Y', how='outer')
-
-    # Renaming column and setting index
-    merged_df = merged_df.rename(columns={'Y': 'DEPT'}).set_index('DEPT')
-
-    return merged_df
